dashboard/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
streamlit-autorefresh>=1.0.1
streamlit-lightweight-charts>=0.7.0
python-dotenv>=1.0.0
yfinance>=0.2.0
